Fuzzy Match Between Two Vectors

This function performs a fuzzy match between two character vectors using various distance methods. It calculates the similarity scores based on the selected method and returns a data frame of matches with their corresponding similarity scores. Optionally, the results can be sorted by the similarity scores.

Usage

fuzzy_match(
  x,
  y,
  method = c("osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw",
    "soundex"),
  useBytes = FALSE,
  weight = c(d = 1, i = 1, s = 1, t = 1),
  q = 1,
  p = 0,
  bt = 0,
  nthread = getOption("sd_num_thread"),
  sorted_by_score = TRUE
)

Arguments

x

A character vector.

y

Another character vector to be matched against `x`.

method

A character string specifying the method to be used for calculating string distance. Valid options are "osa", "lv", "dl", "hamming", "lcs", "qgram", "cosine", "jaccard", "jw", and "soundex".

useBytes

Logical; if `TRUE`, the matching is done byte-by-byte rather than character-by-character.

weight

A named numeric vector of weights to be used for the "osa", "lv", and "dl" methods.

q

An integer used for the "qgram" method.

p

A penalty parameter used in the "jw" method.

bt

A breaking tie parameter used in some methods.

nthread

The number of threads to use for computation.

sorted_by_score

Logical; if `TRUE`, the results are sorted by descending similarity scores.

Value

A `data.frame` containing the original elements of `x` and `y`, their similarity scores, and index positions. If `sorted_by_score` is `TRUE`, the data frame is sorted by the similarity scores in descending order.

Examples

x <- "apple"
y <- c("ape", "bandana", "berry")
fuzzy_match(x, y, method = "jaccard")
#>       x       y similarity_scores index
#> 1 apple     ape         0.7083333     1
#> 2 apple bandana         0.0000000     2
#> 3 apple   berry         0.0000000     3